move toggl functions
